review fixes & change secondsToReachTrack, secondsToReachProject type to Float

+++ shared/src/commonMain/kotlin/org/hyperskill/app/study_plan/domain/model/StudyPlan.kt
@@ -15,9 +15,9 @@ data class StudyPlan(
     @SerialName("sections")
     val sections: List<Long>,
     @SerialName("seconds_to_reach_track")
-    val secondsToReachTrack: Long,
+    val secondsToReachTrack: Float,
     @SerialName("seconds_to_reach_project")
-    val secondsToReachProject: Long,
+    val secondsToReachProject: Float,
     @SerialName("created_at")
     val createdAt: String,
     @SerialName("status")
